Lab3
Write a program that asks the user to enter the number of views the video has received in the last 30 days and the engagement rate for that video. The program should then display the advice based on the rules above and if met, the revenue based on the ad type.