Lab5-1 and 2
1. Write an IPO diagram and Python program that has two functions, main and determine_stars.

main – Should accept input of five numeric ratings from the user USING A LOOP.   It should then calculate the average numeric score for the restaurant.    The numeric average should be passed to the determine_stars function.
 
determine_stars – should display the number of stars based on the numeric average:
 
Greater than 9:   *****
8.0 - 8.9:   ​          ****
7.0 – 7.9:             ***
6.0 – 6.9:             **
5.0 – 5.9:  ​           *
Below 5.0:           No stars

2. When an object is falling because of gravity, the following formula can be used to determine the distance the object falls in a specific time period:
 
d = ½ gt2
 
The variables in the formula are as follows:
d is the distance in meters
g is 9.8 (the gravitational constant)
t is the amount of time in seconds the object has been falling
 
Your program will calculate the distance in meters based on the object’s falling distance.