Lab9
Write a program that creates a dictionary containing course numbers and the room numbers of the rooms where the courses meet. The program should also create a dictionary containing course numbers and the names of the instructors that teach the course. The program should also create a dictionary containing course numbers and the meeting times of each course. The program should let the user enter a course number, then it should display the course's room number, instructor, and meeting time. If the course is not in the dictionary, a message should be displayed. The program should. contain a loop which continues to ask for a course number until the user chooses to quit.

Course Numbers: CS101, CS102, CS103, NT110, CM241